% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_lp.R
\name{as_pp_lp}
\alias{as_pp_lp}
\alias{as_pp_lp.pp_lp}
\alias{as_pp_lp.matrix}
\alias{as_pp_lp.ppp}
\title{Convert an object to a pp_lp object}
\usage{
as_pp_lp(x, ...)

\method{as_pp_lp}{pp_lp}(x, ...)

\method{as_pp_lp}{matrix}(x, domain, bbox = NULL, ...)

\method{as_pp_lp}{ppp}(x, ...)
}
\arguments{
\item{x}{An object to convert. Can be a \code{matrix}, a \code{pp_lp}, or a
\code{spatstat.geom::ppp}.}

\item{...}{Additional arguments (e.g. \code{domain} for the matrix method).}

\item{domain}{A \code{domain_lp} object (required for the matrix method).}

\item{bbox}{A bounding box matrix (optional).}
}
\value{
An object of class \code{"pp_lp"}.
}
\description{
Convert an object to a pp_lp object
}
\examples{
X <- matrix(runif(40), 20, 2)
dom <- domain_Rd(2L)
pp <- as_pp_lp(X, domain = dom)
print(pp)
}
