% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cv_lp_fixed_h_m_cpp}
\alias{cv_lp_fixed_h_m_cpp}
\title{Optimized LOO CV scores for a fixed (m, h) grid point (C++ backend)}
\usage{
cv_lp_fixed_h_m_cpp(X, U_quad_list, n_total_vec, h, alphas)
}
\arguments{
\item{X}{Numeric matrix d x n of all observations.}

\item{U_quad_list}{List of d x N_i matrices of quadrature points.}

\item{n_total_vec}{Integer vector of total draws in the box for each point.
(length n).}

\item{h}{Bandwidth (scalar > 0).}

\item{alphas}{Integer matrix D_m x d of multi-indices.}
}
\value{
Numeric vector of length n: log(f_loo(X_i)).
}
\description{
Optimized LOO CV scores for a fixed (m, h) grid point (C++ backend)
}
\keyword{internal}
