% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{density_lp_point}
\alias{density_lp_point}
\title{Estimate the density at a single point t using local polynomials}
\usage{
density_lp_point(X, t, h, m, domain, N_quad)
}
\arguments{
\item{X}{Numeric matrix n x d of observations.}

\item{t}{Numeric vector of length d (target point).}

\item{h}{Bandwidth (scalar > 0).}

\item{m}{Polynomial degree (integer >= 0).}

\item{domain}{An \code{"domain_lp"} object.}

\item{N_quad}{Number of Monte Carlo quadrature points.}
}
\value{
Scalar estimate of f(t).
}
\description{
Estimate the density at a single point t using local polynomials
}
\keyword{internal}
