% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityLP-package.R
\docType{package}
\name{densityLP-package}
\alias{densityLP}
\alias{densityLP-package}
\title{densityLP: Local Polynomial Density Estimation on Complicated Domains}
\description{
Nonparametric local polynomial estimator for multivariate density functions on known domains of arbitrary dimension, including complicated domains with sharp concavities, holes, and local pinches such as polynomial sectors. Implements the procedure of Bertin, Klutchnikoff and Ouimet (Bernoulli, forthcoming).
}
\author{
\strong{Maintainer}: Nicolas Klutchnikoff \email{nicolas.klutchnikoff@univ-rennes2.fr} (\href{https://orcid.org/0000-0001-9349-0771}{ORCID})

Authors:
\itemize{
  \item Karine Bertin \email{karine.bertin@uv.cl} (\href{https://orcid.org/0000-0002-8167-137X}{ORCID})
  \item Frederic Ouimet \email{frederic.ouimet2@mcgill.ca} (\href{https://orcid.org/0000-0001-7933-5265}{ORCID})
}

}
\keyword{internal}
