% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain_func}
\alias{domain_func}
\title{Domain defined by an indicator function}
\usage{
domain_func(is_in_domain, d)
}
\arguments{
\item{is_in_domain}{Vectorised function \code{f(X_mat)} where \code{X_mat} is an
\verb{n x d} numeric matrix of points in R^d and the return value is a logical
vector of length \code{n}.}

\item{d}{Space dimension (integer >= 1).}
}
\value{
An \code{"lp_domain"} object.
}
\description{
Domain defined by an indicator function
}
