% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain_Rd}
\alias{domain_Rd}
\title{Domain R^d (no constraint: neighbourhood = full \eqn{[-h, h]^d})}
\usage{
domain_Rd(d)
}
\arguments{
\item{d}{Space dimension (integer >= 1).}
}
\value{
An \code{"lp_domain"} object.
}
\description{
Domain R^d (no constraint: neighbourhood = full \eqn{[-h, h]^d})
}
