% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gram_matrix_cpp}
\alias{gram_matrix_cpp}
\title{Compute the Gram matrix B_gamma in C++}
\usage{
gram_matrix_cpp(U, h, alphas, n_total)
}
\arguments{
\item{U}{Numeric matrix d x N of quadrature points centred at t.}

\item{h}{Bandwidth (scalar > 0).}

\item{alphas}{Integer matrix D_m x d of multi-indices.}

\item{n_total}{Total draws in the box \eqn{[-h,h]^d} (from the sampler).}
}
\value{
Symmetric matrix D_m x D_m.
}
\description{
Compute the Gram matrix B_gamma in C++
}
\keyword{internal}
