% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{new_cv_density_lp}
\alias{new_cv_density_lp}
\title{Internal constructor for cv_density_lp objects}
\usage{
new_cv_density_lp(m_hat, h_hat, scores, grids, call)
}
\description{
Internal constructor for cv_density_lp objects
}
\keyword{internal}
