% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_lp.R
\name{new_density_lp}
\alias{new_density_lp}
\title{Internal constructor for density_lp objects}
\usage{
new_density_lp(
  estimate,
  variance,
  t_grid,
  X,
  h,
  m,
  domain,
  N_quad,
  n_fail,
  call
)
}
\description{
Internal constructor for density_lp objects
}
\keyword{internal}
