% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_lp_ppp.R
\name{new_density_lp_ppp}
\alias{new_density_lp_ppp}
\title{Internal constructor for density_lp_ppp objects}
\usage{
new_density_lp_ppp(im_obj, params, stats, call)
}
\description{
Internal constructor for density_lp_ppp objects
}
\keyword{internal}
