% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.density_lp}
\alias{plot.density_lp}
\title{Plot a density_lp object}
\usage{
\method{plot}{density_lp}(
  x,
  main = paste0("densityLP  (h = ", x$params$h, ", m = ", x$params$m, ")"),
  ...
)
}
\arguments{
\item{x}{A \code{"density_lp"} object.}

\item{main}{Plot title (default: bandwidth and degree).}

\item{...}{Passed to the underlying plot function.}
}
\value{
\code{x} invisibly.
}
\description{
Plot a density_lp object
}
