% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.density_lp_ppp}
\alias{plot.density_lp_ppp}
\title{Plot a density_lp_ppp object}
\usage{
\method{plot}{density_lp_ppp}(x, ...)
}
\arguments{
\item{x}{A \code{"density_lp_ppp"} object.}

\item{...}{Passed to \code{\link[spatstat.geom:plot.im]{spatstat.geom::plot.im()}}.}
}
\value{
\code{x} invisibly.
}
\description{
Plot a density_lp_ppp object
}
