% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cv_density_lp}
\alias{print.cv_density_lp}
\title{Print a cv_density_lp object}
\usage{
\method{print}{cv_density_lp}(x, ...)
}
\arguments{
\item{x}{A \code{"cv_density_lp"} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a cv_density_lp object
}
