% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.density_lp}
\alias{print.density_lp}
\title{Print a density_lp object}
\usage{
\method{print}{density_lp}(x, ...)
}
\arguments{
\item{x}{A \code{"density_lp"} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a density_lp object
}
