% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.density_lp_ppp}
\alias{print.density_lp_ppp}
\title{Print a density_lp_ppp object}
\usage{
\method{print}{density_lp_ppp}(x, ...)
}
\arguments{
\item{x}{A \code{"density_lp_ppp"} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a density_lp_ppp object
}
