% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{print.domain_lp}
\alias{print.domain_lp}
\title{Print a domain_lp object}
\usage{
\method{print}{domain_lp}(x, ...)
}
\arguments{
\item{x}{A \code{"domain_lp"} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a domain_lp object
}
