% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{print.lp_domain}
\alias{print.lp_domain}
\title{Print an lp_domain object}
\usage{
\method{print}{lp_domain}(x, ...)
}
\arguments{
\item{x}{An \code{"lp_domain"} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print an lp_domain object
}
