% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_rejection}
\alias{sampler_rejection}
\title{Rejection sampler for an arbitrary analytic domain}
\usage{
sampler_rejection(is_in_domain, n_pre = 2000L)
}
\arguments{
\item{is_in_domain}{Vectorised function: n x d matrix -> logical vector.}

\item{n_pre}{Number of points used for the preliminary acceptance rate
estimate.}
}
\value{
Closure \verb{function(N, t, h)}.
}
\description{
Rejection sampler for an arbitrary analytic domain
}
\keyword{internal}
