% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cv_density_lp}
\alias{summary.cv_density_lp}
\title{Summary of a cv_density_lp object}
\usage{
\method{summary}{cv_density_lp}(object, ...)
}
\arguments{
\item{object}{A \code{"cv_density_lp"} object.}

\item{...}{Ignored.}
}
\value{
\code{object} invisibly.
}
\description{
Summary of a cv_density_lp object
}
