% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.density_lp}
\alias{summary.density_lp}
\title{Summary of a density_lp object}
\usage{
\method{summary}{density_lp}(object, ...)
}
\arguments{
\item{object}{A \code{"density_lp"} object.}

\item{...}{Ignored.}
}
\value{
\code{object} invisibly.
}
\description{
Summary of a density_lp object
}
