% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.density_lp_ppp}
\alias{summary.density_lp_ppp}
\title{Summary of a density_lp_ppp object}
\usage{
\method{summary}{density_lp_ppp}(object, ...)
}
\arguments{
\item{object}{A \code{"density_lp_ppp"} object.}

\item{...}{Ignored.}
}
\value{
\code{object} invisibly.
}
\description{
Summary of a density_lp_ppp object
}
