% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain_from_indicator}
\alias{domain_from_indicator}
\title{Domain defined by an indicator function}
\usage{
domain_from_indicator(is_in_domain)
}
\arguments{
\item{is_in_domain}{Vectorised indicator function of \code{d} arguments (one per
coordinate axis), returning a logical vector of length \code{n}.  The dimension
\code{d} is inferred from \code{length(formals(is_in_domain))}.}
}
\value{
An \code{"domain_lp"} object.
}
\description{
Domain defined by an indicator function
}
\examples{
is_in_domain <- function(x, y) x^2 + y^2 <= 1
dom <- domain_from_indicator(is_in_domain)
print(dom)
}
