% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.density_lp}
\alias{plot.density_lp}
\title{Plot a density_lp object}
\usage{
\method{plot}{density_lp}(x, ...)
}
\arguments{
\item{x}{A \code{"density_lp"} object.}

\item{...}{Passed to the underlying plot function.}
}
\value{
\code{x} invisibly.
}
\description{
Plot a density_lp object
}
